requirements.txt
fastapi
uvicorn
requests
google-generativeai
pydantic
slowapi
httpx
bcrypt
python-multipart
python-jose
sqlalchemy
python-dotenv
pandas
numpy
scikit-learn
ecdsa
gunicorn
